Add script to enable Netcool Integration
Use this script with IBM Cloud Pak for AIOps 4.3.0 and later to allow a
Netcool Operations Insight probe to be connected to it.
The script can be used on version 4.2.x before upgrading to 4.3.0 in order to
ensure uninterrupted operation.

Usage: ./enabled-netcool-integration.sh [-n namespace] [cp4aiops_name]

The current namespace is used if the -n option is not used.
cp4aiops_name defaults to ibm-cp-aiops.
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Signed-off-by: Nick Banyard <banyard@uk.ibm.com>